Remove unnecessary tests in PropertyGroupingTest
