+++ java/configuration/src/main/java/sleeper/configuration/properties/table/TablePropertyGroup.java
@@ -23,7 +23,6 @@ import java.util.Collections;
 import java.util.List;
 
 import static sleeper.configuration.properties.PropertyGroup.group;
-import static sleeper.configuration.properties.PropertyGroup.sortPropertiesByGroup;
 
 public class TablePropertyGroup {
     private TablePropertyGroup() {
@@ -52,7 +51,4 @@ public class TablePropertyGroup {
         return Collections.unmodifiableList(ALL);
     }
 
-    public static List<TableProperty> sortProperties(List<TableProperty> unsorted) {
-        return sortPropertiesByGroup(unsorted, TablePropertyGroup.getAll());
-    }
 }
