+++ java/configuration/src/test/java/sleeper/configuration/properties/PropertyGroupingTest.java
@@ -23,22 +23,15 @@ import java.util.List;
 import java.util.function.Consumer;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.InstanceProperty.getAll;
 import static sleeper.configuration.properties.InstancePropertyGroup.BULK_IMPORT;
 import static sleeper.configuration.properties.InstancePropertyGroup.COMMON;
 import static sleeper.configuration.properties.InstancePropertyGroup.INGEST;
-import static sleeper.configuration.properties.InstancePropertyGroup.sortPropertiesByGroup;
 
-class InstancePropertyGroupingTest {
-    @Test
-    void shouldGetAllUserDefinedAndSystemDefinedProperties() {
-        // Given/When
-        List<InstanceProperty> propertyList = getAll();
+class PropertyGroupingTest {
+    private final List<PropertyGroup> groups = List.of(COMMON, INGEST, BULK_IMPORT);
 
-        // Then
-        assertThat(propertyList)
-                .containsAll(UserDefinedInstanceProperty.getAll())
-                .containsAll(SystemDefinedInstanceProperty.getAll());
+    static {
+        InstanceProperty.getAll();
     }
 
     @Test
@@ -92,7 +85,11 @@ class InstancePropertyGroupingTest {
                 .containsExactly(userProperty1, systemProperty1, userProperty2, systemProperty2, userProperty3, systemProperty3);
     }
 
-    private static UserDefinedInstanceProperty userProperty(
+    private List<InstanceProperty> sortPropertiesByGroup(List<InstanceProperty> propertyList) {
+        return PropertyGroup.sortPropertiesByGroup(propertyList, groups);
+    }
+
+    private static InstanceProperty userProperty(
             String name, PropertyGroup group, Consumer<UserDefinedInstanceProperty> addToList) {
         return UserDefinedInstancePropertyImpl.named(name)
                 .propertyGroup(group)
@@ -100,7 +97,7 @@ class InstancePropertyGroupingTest {
                 .addToAllList(addToList).build();
     }
 
-    private static SystemDefinedInstanceProperty systemProperty(
+    private static InstanceProperty systemProperty(
             String name, PropertyGroup group, Consumer<SystemDefinedInstanceProperty> addToList) {
         return SystemDefinedInstancePropertyImpl.named(name)
                 .propertyGroup(group)
